requirement.txt
lxml==4.6.3
python-docx==0.8.11
mysql-connector-python==8.0.26
